requirements.txt
# Core dependencies
rich>=10.16.2
Pillow>=9.0.0
watchdog>=2.1.9
pyyaml>=6.0
click>=8.0.0
python-dotenv>=0.19.0

# UI and styling
textual>=0.1.18
prompt_toolkit>=3.0.20

# Image processing
numpy>=1.22.0
click
numpy
Pillow
prompt_toolkit
python-dotenv
pyyaml
rich
textual
watchdog
click
Pillow
python-dotenv
pyyaml
rich
watchdog
click>=8.0.0
numpy>=1.22.0
Pillow>=10.2.0
prompt_toolkit>=3.0.20
python-dotenv>=0.19.0
pyyaml>=6.0
rich>=10.16.2
textual>=0.1.18
watchdog>=2.1.9